finish summarize services

+++ src/main/java/com/eventmap/fluent/domain/SummarizeResult.java
@@ -11,7 +11,7 @@ public class SummarizeResult {
 
     private int totalWords;
 
-    private int speechSpeedAverage;
+    private float speechSpeedAverage;
 
     public int getCorrectWords() {
         return correctWords;
@@ -37,11 +37,11 @@ public class SummarizeResult {
         this.totalWords = totalWords;
     }
 
-    public int getSpeechSpeedAverage() {
+    public float getSpeechSpeedAverage() {
         return speechSpeedAverage;
     }
 
-    public void setSpeechSpeedAverage(int speechSpeedAverage) {
+    public void setSpeechSpeedAverage(float speechSpeedAverage) {
         this.speechSpeedAverage = speechSpeedAverage;
     }
 }
